requirements.txt
numpy==1.26.1
pandas==2.0.3
pydeck==0.8.1b0
streamlit==1.32.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.92.0
PyPDF2==3.0.1
gspread
boto3
importlib
